Submit Diagnosys to Server
