+++ Covid19Radar/Covid19Radar.iOS/Services/DeviceCheckService.cs
@@ -16,22 +16,11 @@ namespace Covid19Radar.iOS.Services
 {
     public class DeviceCheckService : IDeviceVerifier
     {
-        /// <summary>
-        /// Verification device information required for positive submissions
-        /// </summary>
-        /// <returns>Device Verification Payload</returns>
 		public async Task<string> VerifyAsync(SelfDiagnosisSubmission submission)
-        {
-            var token = await DeviceCheck.DCDevice.CurrentDevice.GenerateTokenAsync();
-            return Convert.ToBase64String(token.ToArray());
-        }
-/*
-        public async Task<string> GetDeviceVerificationPayload()
-        {
-            if (!DCDevice.CurrentDevice.Supported) return "";
-            var token = await DCDevice.CurrentDevice.GenerateTokenAsync();
-            return token.GetBase64EncodedString(NSDataBase64EncodingOptions.None);
-        }
-*/
-    }
+		{
+			var token = await DeviceCheck.DCDevice.CurrentDevice.GenerateTokenAsync();
+			return Convert.ToBase64String(token.ToArray());
+		}
+
+	}
 }
